Move CancellationToken types back
